Update Gradle settings and add item behavior controls
The gradle settings have been updated to use a dotenv plugin and to publish to GitHubPackages. In addition, item behavior in the game has been enhanced with the ability to control whether they can be picked up, dropped, or broken. Changing the Java project language level was also part of this commit.